some capacity assert logic in book-and-seller-replace-items. Commented out for now. Ill try to come back to this later
