+++ packages/openactive-integration-tests/test/features/criteria-requirements.json
@@ -119,7 +119,7 @@
     },
     "seller-requested-cancellation-message": {
       "primary": {
-        "TestOpportunityBookable": 1
+        "TestOpportunityBookable": 4
       }
     },
     "seller-requested-replacement": {
