+++ packages/openactive-integration-tests/test/shared-behaviours/test-recipes.js
@@ -34,25 +34,33 @@ const TestRecipes = {
     /** @type {RunTestsFn} */
     const runTestsFn = (configuration, orderItemCriteriaList, featureIsImplemented, logger) => {
       // ## Initiate Flow Stages
-      const { fetchOpportunities, c1, c2, bookRecipe, defaultFlowStageParams } = FlowStageRecipes.initialiseSimpleC1C2BookFlow(orderItemCriteriaList, logger);
-      const [testInterfaceAction, orderFeedUpdate] = OrderFeedUpdateFlowStageUtils.wrap({
-        wrappedStageFn: prerequisite => (new TestInterfaceActionFlowStage({
-          ...defaultFlowStageParams,
-          testName: `Test Interface Action (${testInterfaceActionParams.actionType})`,
-          prerequisite,
-          createActionFn: () => ({
-            type: testInterfaceActionParams.actionType,
-            // Note that these 2 fields may need to be configurable in future:
-            objectType: 'Order',
-            objectId: bookRecipe.b.getOutput().orderId,
-          }),
-        })),
-        orderFeedUpdateParams: {
-          ...defaultFlowStageParams,
-          prerequisite: bookRecipe.lastStage,
-          testName: `Orders Feed (after ${testInterfaceActionParams.actionType})`,
-        },
-      });
+      const {
+        fetchOpportunities,
+        c1,
+        c2,
+        bookRecipe,
+        testInterfaceAction,
+        orderFeedUpdate,
+      } = FlowStageRecipes.successfulC1C2BookFollowedByTestInterfaceAction(orderItemCriteriaList, logger, testInterfaceActionParams);
+      // const { fetchOpportunities, c1, c2, bookRecipe, defaultFlowStageParams } = FlowStageRecipes.initialiseSimpleC1C2BookFlow(orderItemCriteriaList, logger);
+      // const [testInterfaceAction, orderFeedUpdate] = OrderFeedUpdateFlowStageUtils.wrap({
+      //   wrappedStageFn: prerequisite => (new TestInterfaceActionFlowStage({
+      //     ...defaultFlowStageParams,
+      //     testName: `Test Interface Action (${testInterfaceActionParams.actionType})`,
+      //     prerequisite,
+      //     createActionFn: () => ({
+      //       type: testInterfaceActionParams.actionType,
+      //       // Note that these 2 fields may need to be configurable in future:
+      //       objectType: 'Order',
+      //       objectId: bookRecipe.b.getOutput().orderId,
+      //     }),
+      //   })),
+      //   orderFeedUpdateParams: {
+      //     ...defaultFlowStageParams,
+      //     prerequisite: bookRecipe.lastStage,
+      //     testName: `Orders Feed (after ${testInterfaceActionParams.actionType})`,
+      //   },
+      // });
 
       // ## Set up tests
       FlowStageUtils.describeRunAndCheckIsSuccessfulAndValid(fetchOpportunities);
